requirements.txt
albumentations==1.3.1
h5py==3.9.0
imageio==2.34.1
ipdb==0.13.13
matplotlib==3.7.2
mlflow==2.8.0
nibabel==5.2.1
numpy==1.24.1
opencv_python==4.9.0.80
packaging==23.1
pandas==2.0.3
Pillow==10.4.0
ptflops==0.7
PyYAML==6.0.2
scikit_learn==1.3.0
scipy==1.14.1
segmentation_models_pytorch==0.3.3
SimpleITK==2.4.0
scikit-image==0.24
tensorboardX==2.6.2.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
ffmpeg-python